requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
scikit-learn>=1.0.0
Pillow>=8.2.0
tqdm>=4.60.0
PyQt5>=5.15.0 